+++ src/main/java/by/it_academy/jd2/Mk_JD2_98_23/controllers/filters/EncodingFilter.java
@@ -1,4 +1,4 @@
-package by.it_academy.jd2.Mk_JD2_98_23.web;
+package by.it_academy.jd2.Mk_JD2_98_23.controllers.filters;
 
 import jakarta.servlet.*;
 import jakarta.servlet.annotation.WebFilter;
