+++ src/main/java/by/it_academy/jd2/Mk_JD2_98_23/controllers/web/MessageServlet.java
@@ -1,4 +1,4 @@
-package by.it_academy.jd2.Mk_JD2_98_23.web;
+package by.it_academy.jd2.Mk_JD2_98_23.controllers.web;
 
 import by.it_academy.jd2.Mk_JD2_98_23.core.dto.MessageCreateDTO;
 import by.it_academy.jd2.Mk_JD2_98_23.core.dto.MessageDTO;
