+++ src/main/webapp/WEB-INF/web.xml
@@ -8,7 +8,7 @@
     <display-name>Mk-JD2-98-23</display-name>
     <filter>
         <filter-name>EncodingFilter</filter-name>
-        <filter-class>by.it_academy.jd2.Mk_JD2_98_23.web.EncodingFilter</filter-class>
+        <filter-class>by.it_academy.jd2.Mk_JD2_98_23.controllers.filters.EncodingFilter</filter-class>
         <init-param>
             <param-name>encoding</param-name>
             <param-value>UTF-8</param-value>
